kernel: move module handling to modules.c
